refactor(userBook): domain/user/userbook 을 domain/userBook 최상위 도메인으로 승격
