+++ src/main/java/com/readum/domain/userBook/dto/UserBookCreateCommand.java
@@ -1,4 +1,4 @@
-package com.readum.domain.user.userbook.dto;
+package com.readum.domain.userBook.dto;
 
 public record UserBookCreateCommand(
         Long userId,
