+++ src/main/java/com/readum/domain/userBook/dto/UserBookSearchItemResult.java
@@ -1,4 +1,4 @@
-package com.readum.domain.user.userbook.dto;
+package com.readum.domain.userBook.dto;
 
 import com.readum.model.userBook.repository.projection.UserBookListItemProjection;
