+++ src/main/java/com/readum/presentation/controller/user/dto/UserBookListResponse.java
@@ -1,6 +1,6 @@
 package com.readum.presentation.controller.user.dto;
 
-import com.readum.domain.user.userbook.dto.UserBookSearchResult;
+import com.readum.domain.userBook.dto.UserBookSearchResult;
 
 import java.util.List;
