+++ src/test/java/com/readum/domain/userBook/service/UserBookDeleteServiceCascadeTest.java
@@ -1,7 +1,7 @@
-package com.readum.domain.user.userbook.service;
+package com.readum.domain.userBook.service;
 
-import com.readum.domain.user.userbook.dto.UserBookDeleteCommand;
-import com.readum.domain.user.userbook.dto.UserBookDeleteResult;
+import com.readum.domain.userBook.dto.UserBookDeleteCommand;
+import com.readum.domain.userBook.dto.UserBookDeleteResult;
 import com.readum.model.aiChat.entity.AiChatMessage;
 import com.readum.model.aiChat.entity.AiChatSession;
 import com.readum.model.aiChat.repository.AiChatMessageRepository;
