+++ src/test/java/com/readum/domain/userBook/service/UserBookSearchServiceTest.java
@@ -1,6 +1,6 @@
-package com.readum.domain.user.userbook.service;
+package com.readum.domain.userBook.service;
 
-import com.readum.domain.user.userbook.dto.UserBookSearchResult;
+import com.readum.domain.userBook.dto.UserBookSearchResult;
 import com.readum.model.userBook.repository.UserBookRepository;
 import com.readum.model.userBook.repository.projection.UserBookListItemProjection;
 import org.junit.jupiter.api.Test;
