+++ src/test/java/com/readum/model/userBook/repository/UserBookConcurrencyTest.java
@@ -3,9 +3,9 @@ package com.readum.model.userBook.repository;
 import com.readum.domain.book.dto.BookResult;
 import com.readum.domain.book.out.BookLookupClient;
 import com.readum.domain.exception.ConflictException;
-import com.readum.domain.user.userbook.dto.UserBookCreateCommand;
-import com.readum.domain.user.userbook.dto.UserBookCreateResult;
-import com.readum.domain.user.userbook.service.UserBookCreateService;
+import com.readum.domain.userBook.dto.UserBookCreateCommand;
+import com.readum.domain.userBook.dto.UserBookCreateResult;
+import com.readum.domain.userBook.service.UserBookCreateService;
 import com.readum.model.book.repository.BookRepository;
 import com.readum.model.user.entity.User;
 import com.readum.model.user.repository.UserRepository;
