+++ src/test/java/com/readum/presentation/controller/user/UserBookControllerTest.java
@@ -3,14 +3,14 @@ package com.readum.presentation.controller.user;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.readum.domain.exception.ConflictException;
 import com.readum.domain.exception.NotFoundException;
-import com.readum.domain.user.userbook.dto.UserBookCreateResult;
-import com.readum.domain.user.userbook.dto.UserBookDeleteResult;
-import com.readum.domain.user.userbook.dto.UserBookSearchItemResult;
-import com.readum.domain.user.userbook.dto.UserBookSearchResult;
-import com.readum.domain.user.userbook.exception.UserBookErrorCode;
-import com.readum.domain.user.userbook.service.UserBookCreateService;
-import com.readum.domain.user.userbook.service.UserBookDeleteService;
-import com.readum.domain.user.userbook.service.UserBookSearchService;
+import com.readum.domain.userBook.dto.UserBookCreateResult;
+import com.readum.domain.userBook.dto.UserBookDeleteResult;
+import com.readum.domain.userBook.dto.UserBookSearchItemResult;
+import com.readum.domain.userBook.dto.UserBookSearchResult;
+import com.readum.domain.userBook.exception.UserBookErrorCode;
+import com.readum.domain.userBook.service.UserBookCreateService;
+import com.readum.domain.userBook.service.UserBookDeleteService;
+import com.readum.domain.userBook.service.UserBookSearchService;
 import com.readum.presentation.common.GlobalExceptionHandler;
 import com.readum.presentation.common.security.AuthenticatedUserIdArgumentResolver;
 import com.readum.presentation.controller.user.dto.UserBookCreateRequest;
